requirements.txt
# requirements.txt
crawl4ai==0.4.247
beautifulsoup4
mysql-connector-python
